Add licenses and refactor file structure
