+++ wirecardpaymentgateway/controllers/front/payment.php
@@ -42,8 +42,8 @@ use Wirecard\PaymentSdk\TransactionService;
 use WirecardEE\Prestashop\Helper\OrderManager;
 use WirecardEE\Prestashop\Helper\Logger as WirecardLogger;
 use WirecardEE\Prestashop\Helper\TransactionBuilder;
-use WirecardEE\Prestashop\Classes\Config\Factories\PaymentConfigurationFactory;
-use WirecardEE\Prestashop\Classes\Config\Services\ShopConfigurationService;
+use WirecardEE\Prestashop\Classes\Config\PaymentConfigurationFactory;
+use WirecardEE\Prestashop\Helper\Services\ShopConfigurationService;
 
 /**
  * Class WirecardPaymentGatewayPaymentModuleFrontController
