+++ wirecardpaymentgateway/controllers/front/return.php
@@ -39,8 +39,8 @@ use Wirecard\PaymentSdk\TransactionService;
 use WirecardEE\Prestashop\Helper\OrderManager;
 use WirecardEE\Prestashop\Helper\Logger as WirecardLogger;
 use WirecardEE\Prestashop\Models\PaymentPoiPia;
-use WirecardEE\Prestashop\Classes\Config\Services\ShopConfigurationService;
-use WirecardEE\Prestashop\Classes\Config\Factories\PaymentConfigurationFactory;
+use WirecardEE\Prestashop\Helper\Services\ShopConfigurationService;
+use WirecardEE\Prestashop\Classes\Config\PaymentConfigurationFactory;
 
 /**
  * Class WirecardPaymentGatewayReturnModuleFrontController
